+++ src/configUI.ts
@@ -31,7 +31,8 @@ function getConfigs<T>(Configs: vscode.WorkspaceConfiguration, section: TConfigK
 }
 
 let oldCustomizeHoverFunctionDocStyle: 1 | 2 | null = null;
-let oldSignatureHelpInformation: 0 | 1 | 2 | 3 | null = null;
+let oldSignatureHelpInformation: 1 | 2 | 3 | null = null;
+let oldSignatureHelpInsertTypeInformation: boolean | null = null;
 let oldFilesTryParserIncludeOpt: 'auto' | 'close' | 'open' | null = null;
 
 function getConfig(Configs: vscode.WorkspaceConfiguration): TConfigs {
@@ -104,7 +105,8 @@ function getConfig(Configs: vscode.WorkspaceConfiguration): TConfigs {
             statusBarDisplayColor: getConfigs<string>(Configs, 'AhkNekoHelp.customize.statusBarDisplayColor'),
             CodeAction2GotoDefRef: getConfigs<boolean>(Configs, 'AhkNekoHelp.customize.CodeAction2GotoDefRef'),
             HoverFunctionDocStyle: getConfigs<1 | 2>(Configs, 'AhkNekoHelp.customize.HoverFunctionDocStyle'),
-            signatureHelp: getConfigs<0 | 1 | 2>(Configs, 'AhkNekoHelp.customize.signatureHelp'),
+            signatureHelp: getConfigs<1 | 2 | 3>(Configs, 'AhkNekoHelp.customize.signatureHelp'),
+            signatureHelpInsertType: getConfigs<boolean>(Configs, 'AhkNekoHelp.customize.signatureHelpInsertType'),
             displayLogMessage: getConfigs<string>(Configs, 'AhkNekoHelp.customize.displayLogMessage'),
         },
         RenameFunctionInStr: getConfigs<boolean>(Configs, 'AhkNekoHelp.Rename.functionInStr'),
@@ -129,6 +131,15 @@ function getConfig(Configs: vscode.WorkspaceConfiguration): TConfigs {
         oldSignatureHelpInformation = ed.customize.signatureHelp;
         void vscode.window.showWarningMessage(
             'Configs change: please restart vscode!\n\n ("AhkNekoHelp.customize.signatureHelp")',
+        );
+    }
+
+    if (oldSignatureHelpInsertTypeInformation === null) {
+        oldSignatureHelpInsertTypeInformation = ed.customize.signatureHelpInsertType;
+    } else if (oldSignatureHelpInsertTypeInformation !== ed.customize.signatureHelpInsertType) {
+        oldSignatureHelpInsertTypeInformation = ed.customize.signatureHelpInsertType;
+        void vscode.window.showWarningMessage(
+            'Configs change: please restart vscode!\n\n ("AhkNekoHelp.customize.signatureHelpInsertType")',
         );
     }