+++ src/provider/SignatureHelpProvider/SignatureFuncUser.ts
@@ -1,10 +1,13 @@
-/* eslint-disable max-depth */
-/* eslint-disable no-magic-numbers */
+/* eslint no-magic-numbers: ["error", { "ignore": [-1,0,1,2,3] }] */
 import * as vscode from 'vscode';
-import type { CAhkFunc, TParamMetaOut } from '../../AhkSymbol/CAhkFunc';
+import type {
+    CAhkFunc,
+    TFnParamMeta,
+    TParamMetaOut,
+} from '../../AhkSymbol/CAhkFunc';
 import { getCustomize } from '../../configUI';
+import type { DeepReadonly } from '../../globalEnum';
 import { CMemo } from '../../tools/CMemo';
-import { getFuncAhkDocData } from '../../tools/Func/getFuncAhkDocData';
 import { getLStr } from '../../tools/str/removeSpecialChar';
 import { ToUpCase } from '../../tools/str/ToUpCase';
 import type { TFnSignData } from './TFnSignData';
@@ -24,46 +27,132 @@ function SignatureFuncSnipText2Sign(selectionRangeText: string): string {
     return arr1.join('\n').trim();
 }
 
-const UserFuncSign = new CMemo<CAhkFunc, TFnSignData>(
-    (fnSymbol: CAhkFunc): TFnSignData => {
-        const {
-            paramMap,
-            selectionRangeText,
-            name,
-            meta,
-        } = fnSymbol;
+type TParamMetaMap = ReadonlyMap<string, TFnParamMeta>;
+function fixParamWithTypeStr(k: string, ParamMetaOut: TParamMetaOut, paramMetaMap: TParamMetaMap): string {
+    const { keyRawName, isByRef, isVariadic } = ParamMetaOut;
+
+    const p1: string = isVariadic
+        ? `${keyRawName}*`
+        : keyRawName;
+
+    const p2: string = isByRef
+        ? `ByRef ${p1}`
+        : p1;
+
+    const paramType: TFnParamMeta | undefined = paramMetaMap.get(k);
+    return paramType === undefined || paramType.ATypeDef === ''
+        ? p2
+        : `${p2}: ${paramType.ATypeDef}`;
+}
+
+function getLabelWithType(
+    InsertType: boolean,
+    style: 0 | 1 | 2 | 3,
+    fnSymbol: CAhkFunc,
+): { parameters: vscode.ParameterInformation[], label: string } {
+    const {
+        paramMap,
+        name,
+        selectionRangeText,
+        meta,
+    } = fnSymbol;
+
+    const { ahkDocMeta, returnList } = meta;
+    const { paramMeta, returnMeta } = ahkDocMeta;
+
+    const paramMetaMap: TParamMetaMap = new Map(paramMeta
+        .map((v: TFnParamMeta): [string, DeepReadonly<TFnParamMeta>] => [ToUpCase(v.BParamName), v]));
+
+    if (!InsertType) {
+        const parameters: vscode.ParameterInformation[] = [];
+        for (const [k, { keyRawName }] of paramMap) {
+            const paramLabel: string = keyRawName;
+
+            const paramType: TFnParamMeta | undefined = style > 0
+                ? paramMetaMap.get(k)
+                : undefined;
+
+            const documentationFix: string | undefined = paramType === undefined
+                ? undefined
+                : paramType.CInfo.join('\n');
+            parameters.push(
+                new vscode.ParameterInformation(
+                    paramLabel,
+                    documentationFix,
+                ),
+            );
+        }
+        return {
+            parameters,
+            label: SignatureFuncSnipText2Sign(selectionRangeText),
+        };
+    }
+
+    const parameters: vscode.ParameterInformation[] = [];
+    const paramList: string[] = [];
+    for (const [k, ParamMetaOut] of paramMap) {
+        const paramLabel: string = fixParamWithTypeStr(k, ParamMetaOut, paramMetaMap);
+        paramList.push(paramLabel);
+
+        const documentation: string | undefined = style > 0
+            ? paramMetaMap.get(k)?.CInfo.join('\n')
+            : undefined;
+
+        parameters.push(
+            new vscode.ParameterInformation(
+                paramLabel,
+                documentation,
+            ),
+        );
+    }
+
+    const returnType: string = returnList.length > 0 || returnMeta.typeDef !== ''
+        ? `: ${
+            returnMeta.typeDef === ''
+                ? 'unknown'
+                : returnMeta.typeDef
+        }`
+        : '';
+
+    return {
+        parameters,
+        label: `${name}(${paramList.join(', ')})${returnType}`,
+    };
+}
+
+function UserFuncSignCore(fnSymbol: CAhkFunc): TFnSignData {
+    const {
+        paramMap,
+        name,
+        meta,
+    } = fnSymbol;
+
+    const style: 0 | 1 | 2 | 3 = getCustomize().signatureHelp;
+    const InsertType: boolean = getCustomize().signatureHelpInsertType;
+
+    const paramList: TParamMetaOut[] = [...paramMap.values()];
+
+    const { length } = paramList;
+    const lastIsVariadic: boolean = paramList.at(-1)?.isVariadic ?? false;
+
+    const { parameters, label } = getLabelWithType(InsertType, style, fnSymbol);
+
+    const doc: vscode.MarkdownString = new vscode.MarkdownString('', true);
+    doc.supportHtml = true;
+
+    if (style >= 2) {
         const { returnList, ahkDocMeta } = meta;
 
-        const { paramSignMap, returnSign } = getFuncAhkDocData(fnSymbol);
-
-        // eslint-disable-next-line no-magic-numbers
-        const style: 0 | 1 | 2 | 3 = getCustomize().signatureHelp;
-
-        const paramList: TParamMetaOut[] = [...paramMap.values()];
-        const parameters: vscode.ParameterInformation[] = paramList
-            .map(({ keyRawName }: TParamMetaOut): vscode.ParameterInformation => (
-                new vscode.ParameterInformation(
-                    keyRawName,
-                    style > 0
-                        ? paramSignMap.get(ToUpCase(keyRawName))
-                        : undefined,
-                )
-            ));
-
-        const { length } = paramList;
-        const last: TParamMetaOut | undefined = paramList.at(-1);
-        const lastIsVariadic: boolean = last?.isVariadic ?? false;
-        const label0: string = selectionRangeText.includes('\n')
-            ? SignatureFuncSnipText2Sign(selectionRangeText)
-            : selectionRangeText;
-
-        const label1: string = returnSign === '' || style > 0
-            ? label0
-            : `${returnSign} := ${label0}`;
-
-        const doc: vscode.MarkdownString = new vscode.MarkdownString('', true);
-
-        if (style >= 2) {
+        const { info, typeDef } = ahkDocMeta.returnMeta;
+        if (info.length > 0 || typeDef !== '') {
+            if (info.length > 0 && typeDef !== '') {
+                doc.appendMarkdown(
+                    InsertType
+                        ? `_@returns_ - ${info.join('\n')}`
+                        : `_@returns_ {${typeDef}} - ${info.join('\n')}`,
+                );
+            }
+        } else {
             doc.appendCodeblock(
                 `${name}(...)${
                     getCustomize().HoverFunctionDocStyle === 1
@@ -75,25 +164,36 @@ const UserFuncSign = new CMemo<CAhkFunc, TFnSignData>(
                 }\n}`,
                 'ahk',
             );
-
-            const { otherMeta } = ahkDocMeta;
-            if (style === 3 && otherMeta.length > 0) {
-                doc
-                    .appendMarkdown('\n***\n')
-                    .appendMarkdown(otherMeta.join('\n'));
-            }
         }
 
-        const SignInfo = new vscode.SignatureInformation(label1, doc);
-        SignInfo.parameters = parameters;
+        const { otherMeta } = ahkDocMeta;
+        if (style === 3) {
+            doc.appendMarkdown(
+                otherMeta
+                    .map((v: string): string => {
+                        if ((/^[ \t]*@\w+[ \t]/u).test(v)) {
+                            return v.replace(
+                                /^[ \t]*@\w+[ \t]/u,
+                                (match: string): string => `_${match.trim()}_ — `,
+                            );
+                        }
+                        return v;
+                    }).join('\n'),
+            );
+        }
+    }
 
-        return {
-            SignInfo,
-            paramLength: length,
-            lastIsVariadic,
-        };
-    },
-);
+    const SignInfo = new vscode.SignatureInformation(label, doc);
+    SignInfo.parameters = parameters;
+
+    return {
+        SignInfo,
+        paramLength: length,
+        lastIsVariadic,
+    };
+}
+
+const UserFuncSign = new CMemo<CAhkFunc, TFnSignData>(UserFuncSignCore);
 
 export function SignUserFn(fnSymbol: CAhkFunc, comma: number): vscode.SignatureHelp | null {
     const { SignInfo, paramLength, lastIsVariadic } = UserFuncSign.up(fnSymbol);